Added spotify_game.html and spotify_game.js
Modified spotify_game.html to move javascript code to separate file (spotify_game.js). Altered structure of the game to be more similar to the trivia game for easier integration with backend later. Also altered design of the page so it uses the same style.css page as the other pages for design consistency.